requirements.txt
fastapi[standard]
nats-py
pyjwt
pwdlib[argon2]
pydantic
ipython